Aliexpres sponsored item filter
There is an new annoying "sponsored item" in the Aliexpress results-list that is most of the time not what you're looking for. I.e the sponsored item doesn't have to match your filters you set when searching on Aliexpress (it could be more expensive than the max price you set, or have paid shipping when you turned on free shipping etc) In general it's quite confusing.

This filter blocks the sponsored items only, I didn't know where else to post it. Sorry if this is the wrong way, this is the first time I submit a filter.