Update NanoAdblocker/NanoFilters#118
Can't use the usual methods against this script of aborting on `encodeURIComponent` or a csp on this site because some video palyers depend on `encodeURIComponent` and including all the video players in a csp would be nearly impossible. This fix is an ugly hack, so a neater one would be nice.