block medium.com "pardon the interruption" modal
The filter is aggressive, but this was the only way to target the correct div. The class on the div is randomized in an effort combat adblocking. An alternative approach would be to target a particular string, but that is also prone to stop working the moment the copy inside the modal changes.