fix(7062): Unbreak viki player
Exception: pingAbFactor() is not a function

In JavaScript null is not a function, aside from that, whatever we
return must be truthy. So let's just return a function that returns
true.

Because of the thrown exception, the player plays about 1 second of
content in a continious loop, even if one has a paid account that
disabled ads.