Add a new loadout.tf filter
Previosuly, a random gray square would randomly appear to me. I think it is a leftover from the ad that was removed. Anyway, I am using this filter for more than three months and it did not cause any problems at all, so I think it is a good idea to add my filter to the global filter list.