remove from badlists
https://blog.readthedocs.com/ads-and-adblocking/:
"Nagging users into allowing ads on Read the Docs. We are envisioning more of a polite nag in a similar vein to jsfiddle rather than an “adblock wall” which prevents usage of Read the Docs until it’s allowed. While we could simply change our CSS and ad API to avoid blocking since we host our ads ourselves, we decided not to engage in a cat and mouse game since this work would not benefit users."