Use diff-path hash to encode diff-block name
When encoding the name as the hash part of the path
(i.e. `#...`), there is no need for a `! Diff-Name: ...`
field, and no need for client-side code to extract and
keep track of a `! Diff-Name: ...` field.

The client-side coe which applies patches will look up
the diff-block name from the `! Diff-Path:` field.