Add Parsiga filter list
Parsiga is an illegal bet script/panel named that scamming Iranian people and do money laundering for several years. Its owner Sa’dollah Amirshaghaghi arrested in Turkey in 2024-03-15. This filter probably blocks over 2500 sites that using the panel or related to his mafia.

References:
https://www.iranintl.com/en/202403162425
https://t.me/webamoozir
https://cyberno.ir/page/posts/138/سایت-های-قمار-و-شرط-بندی-ایرانی-از-چه-اسکریپت-هایی-استفاده-می-کنند؟

List of websites is taken from:
https://urlscan.io/search/#githob.co

Also the first line probably blocks the page if user wanted to login to the websites.